+++ src/frontend/test/admin_invitations_page_test.dart
@@ -77,9 +77,11 @@ http.Client _mockClient(
           'email': 'admin@example.com',
           'permissions': {
             '/admin': ['*'],
-            '/admin/users': ['view'],
+            '/admin/users': ['view-users'],
             '/admin/groups': ['view'],
-            '/admin/invitations': ['view'],
+            '/admin/invitations': ['view-invitations'],
+            '/admin/server': ['view-server-schedule'],
+            '/admin/acl': ['change-acls'],
           },
           'groups': [
             {'id': 'g1', 'name': 'admin'}